Moving charting classes to their own cs file
