+++ Common/Series.cs
@@ -15,114 +15,15 @@
 
 using System;
 using System.Collections.Generic;
-using Newtonsoft.Json;
-using QuantConnect.Logging;
 using System.Drawing;
-using Newtonsoft.Json.Converters;
 using System.Runtime.Serialization;
+using Newtonsoft.Json;
+using Newtonsoft.Json.Converters;
+using QuantConnect.Logging;
 using QuantConnect.Util;
 
 namespace QuantConnect
 {
-    /// <summary>
-    /// Single Parent Chart Object for Custom Charting
-    /// </summary>
-    [JsonObject]
-    public class Chart
-    {
-        /// Name of the Chart:
-        public string Name = "";
-
-        /// Type of the Chart, Overlayed or Stacked.
-        [Obsolete("ChartType is now obsolete. Please use Series indexes instead by setting index in the series constructor.")]
-        public ChartType ChartType = ChartType.Overlay;
-
-        /// List of Series Objects for this Chart:
-        public Dictionary<string, Series> Series = new Dictionary<string, Series>();
-
-        /// <summary>
-        /// Default constructor for chart:
-        /// </summary>
-        public Chart() { }
-
-        /// <summary>
-        /// Chart Constructor:
-        /// </summary>
-        /// <param name="name">Name of the Chart</param>
-        /// <param name="type"> Type of the chart</param>
-        [Obsolete("ChartType is now obsolete and ignored in charting. Please use Series indexes instead by setting index in the series constructor.")]
-        public Chart(string name, ChartType type = ChartType.Overlay)
-        {
-            Name = name;
-            Series = new Dictionary<string, Series>();
-            ChartType = type;
-        }
-
-        /// <summary>
-        /// Constructor for a chart
-        /// </summary>
-        /// <param name="name">String name of the chart</param>
-        public Chart(string name)
-        {
-            Name = name;
-            Series = new Dictionary<string, Series>();
-        }
-
-        /// <summary>
-        /// Add a reference to this chart series:
-        /// </summary>
-        /// <param name="series">Chart series class object</param>
-        public void AddSeries(Series series)
-        {
-            //If we dont already have this series, add to the chrt:
-            if (!Series.ContainsKey(series.Name))
-            {
-                Series.Add(series.Name, series);
-            }
-            else
-            {
-                throw new Exception("Chart.AddSeries(): Chart series name already exists");
-            }
-        }
-
-        /// <summary>
-        /// Fetch the updates of the chart, and save the index position.
-        /// </summary>
-        /// <returns></returns>
-        public Chart GetUpdates()
-        {
-            var copy = new Chart(Name);
-            try
-            {
-                foreach (var series in Series.Values)
-                {
-                    copy.AddSeries(series.GetUpdates());
-                }
-            }
-            catch (Exception err) {
-                Log.Error(err);
-            }
-            return copy;
-        }
-
-        /// <summary>
-        /// Return a new instance clone of this object
-        /// </summary>
-        /// <returns></returns>
-        public Chart Clone()
-        {
-            var chart = new Chart(Name);
-
-            foreach (var kvp in Series)
-            {
-                chart.Series.Add(kvp.Key, kvp.Value.Clone());
-            }
-
-            return chart;
-        }
-    }
-
-
     /// <summary>
     /// Chart Series Object - Series data and properties for a chart:
     /// </summary>
@@ -338,7 +239,8 @@ namespace QuantConnect
                 //Shuffle the update point to now:
                 _updatePosition = Values.Count;
             }
-            catch (Exception err) {
+            catch (Exception err)
+            {
                 Log.Error(err);
             }
             return copy;
@@ -379,58 +281,6 @@ namespace QuantConnect
         }
     }
 
-
-    /// <summary>
-    /// Single Chart Point Value Type for QCAlgorithm.Plot();
-    /// </summary>
-    [JsonObject]
-    public class ChartPoint
-    {
-        /// Time of this chart point: lower case for javascript encoding simplicty
-        public long x;
-
-        /// Value of this chart point:  lower case for javascript encoding simplicty
-        public decimal y;
-
-        /// <summary>
-        /// Default constructor. Using in SeriesSampler.
-        /// </summary>
-        public ChartPoint() {}
-
-        /// <summary>
-        /// Constructor that takes both x, y value paris
-        /// </summary>
-        /// <param name="xValue">X value often representing a time in seconds</param>
-        /// <param name="yValue">Y value</param>
-        public ChartPoint(long xValue, decimal yValue)
-        {
-            x = xValue;
-            y = yValue;
-        }
-
-        ///Constructor for datetime-value arguements:
-        public ChartPoint(DateTime time, decimal value)
-        {
-            x = Convert.ToInt64(Time.DateTimeToUnixTimeStamp(time.ToUniversalTime()));
-            y = value.SmartRounding();
-        }
-
-        ///Cloner Constructor:
-        public ChartPoint(ChartPoint point)
-        {
-            x = point.x;
-            y = point.y.SmartRounding();
-        }
-
-        /// <summary>
-        /// Provides a readable string representation of this instance.
-        /// </summary>
-        public override string ToString()
-        {
-            return Time.UnixTimeStampToDateTime(x).ToString("o") + " - " + y;
-        }
-    }
-
     /// <summary>
     /// Available types of charts
     /// </summary>
@@ -450,17 +300,6 @@ namespace QuantConnect
         StackedArea,
         /// Pie chart
         Pie
-    }
-
-    /// <summary>
-    /// Type of chart - should we draw the series as overlayed or stacked
-    /// </summary>
-    public enum ChartType
-    {
-        /// Overlayed stacked
-        Overlay,
-        /// Stacked series on top of each other.
-        Stacked
     }
 
     /// <summary>
